Network Intrusion Detection System
This was a personal endeavour I chose to program entirely in Python. When testing the program, I opened a VM on my machine in order to safely test the code. The goal of this program was to sniff packets over a network, and if the number of packets exceeded a set threshold I set, an automatic email alert would be sent to me.

I hid the password login for security reasons.